Return information about extracted constants from `pg_query_normalize`
`PgQueryNormalizeResult` now includes information about the extracted
constants, their location and extent, as well as the lexer token type
and the value of constants as interpreted by the lexer.  This makes
`pg_query_normalize` usable not just for query identification, but also
as a basis for automatic conversion of literal queries into
constant-agnostic prepared statements or other applications where
auto-parametrization of queries is useful.

This also makes `pg_query_normalize` to use `base_yylex` instead of
`core_yylex` which normalizes `USCONST` constants properly.